added simple user profiles
Lemontrac now allows for editing of user profiles and setting up custom avatars for users. "Profiles" consist of an email address, password and avatar which can be changed and a username which is chosen during sign up.

Other fields may be added at a later date, but nothing has been decided yet.